:pencil2: rename: Rename Radio to Toggle
